# Conflicts:
#	static/js/common.js

+++ engine/app/Http/Controllers/WorkController.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
+use RecycleArt\Models\Tags;
+use RecycleArt\Models\TagsRel;
 use RecycleArt\Models\Work;
 use RecycleArt\Models\WorkImages;
 
@@ -48,11 +50,13 @@ class WorkController extends Controller
     }
 
     /**
+     * Process for add/edit work
+     *
      * @param Request $request
      *
      * @return mixed
      */
-    public function addProcess(Request $request)
+    public function process(Request $request)
     {
         $workId = $request->post('workId') ?: 0;
         $workId = Work::getInstance()->updateOrSave($workId, [
@@ -60,13 +64,23 @@ class WorkController extends Controller
             'description' => $request->post('description'),
             'userId'      => Auth::id(),
         ]);
-        $isSaved = false;
+        $isSaved = (bool)$workId;
         if (!empty($request->file('images'))) {
             $isSaved = WorkImages::getInstance()->addImamges($request->file('images'), $workId);
         }
-        if (!$isSaved) {
-            $request->session()->flash('addWorkResult', __('work.addProcessError'));
-            return Redirect::to('/cabinet/work/new');
+        if (!empty($request->post('tags'))) {
+            $tagsArray = \explode(',', $request->post('tags'));
+            foreach ($tagsArray as $tag) {
+                $tag = \trim($tag, ' ');
+                if (empty($tag)) {
+                    continue;
+                }
+                $existingTag = (new Tags())->getByName($tag);
+                if (empty($existingTag)) {
+                    $existingTag['id'] = (new Tags())->add($tag);
+                }
+                (new TagsRel())->addToWork($existingTag['id'], $workId);
+            }
         }
         $request->session()->flash('addWorkResult', __('work.addProcessSuccess'));
         return Redirect::to('/cabinet/work');
@@ -80,9 +94,8 @@ class WorkController extends Controller
      */
     public function remove(Request $request, int $id)
     {
-        $user = Auth::user();
-        $workPath = public_path('uploads/' . $user->id . '/work/' . $id);
-        if (Work::getInstance()->removeById($id) && WorkImages::getInstance()->removeByWorkId($id)) {
+        $workPath = public_path('uploads/' . Auth::id() . '/work/' . $id);
+        if (Work::getInstance()->removeById($id) && WorkImages::getInstance()->removeByWorkId($id) && (new TagsRel())->deleteByWork($id)) {
             File::cleanDirectory($workPath);
             rmdir($workPath);
             $request->session()->flash('addWorkResult', __('work.addWorkRemovedSuccess'));
@@ -117,17 +130,11 @@ class WorkController extends Controller
      */
     public function edit(int $id)
     {
-        $work = Work::find($id);
+        $work = (new Work)->getById($id);
         if (empty($work)) {
             abort(404, __('workNotFound'));
         }
-        $images = WorkImages::getbyWorkId($id);
-        if (empty($images)) {
-            $images = [];
-        } else {
-            $images = $images->toArray();
-        }
-        return \view('work.form', ['work' => $work->toArray(), 'images' => $images]);
+        return \view('work.form', ['work' => $work]);
     }
 
     /**
